move ccc-lib/http to typescript
also, allow tracing the http requests
